+++ NEWS.md
@@ -4,5 +4,6 @@ Breaking changes
 * `names_section` parameter is replaced by a named vector in `pattern`
 
 Other changes
+* `create_vignette_last_modif()` creates a vignette with dates of files modification in git
 * Added summary tables for each section
 * Added a `NEWS.md` file to track changes to the package.